Reworked voxel break for batched voxel falling
Replaces per-voxel break event with a batched event system in ATVoxelTree, introducing FBreakVoxelsEventSignature and related data structures. Swaps FillChunkWithVoxels and BreakAllVoxelsAtChunk implementations for correct functionality. Updates time budget variables to use milliseconds instead of seconds. Refactors material override application in UATVoxelISMC to use a function call. Updates collision profile for FallingVoxel in DefaultEngine.ini. Removes BP_BrokenVoxel and adds BP_BrokenVoxelChunk asset. Various Blueprint and asset updates to support new voxel breaking logic.

+++ Source/AvalancheTest/Simulation/ATSimulationTask_Avalanche.h
@@ -32,7 +32,7 @@ protected:
 protected:
 
 	UPROPERTY(Category = "Avalanche", EditAnywhere, BlueprintReadWrite)
-	float AvalancheCooldown;
+	double AvalancheCooldown;
 
 	UPROPERTY(Category = "Avalanche", EditAnywhere, BlueprintReadWrite)
 	float InstantAvalancheStabilityThreshold;
